+++ src/components/sign-in/index.ts
@@ -1,0 +1,1 @@
+export { SigninFormInner } from "./sign-in-form-inner";