+++ src/lib/actions/booking/email/index.ts
@@ -1,1 +1,1 @@
-export { sendBookingEmail } from "./send-mail";
+export { sendBookingEmail, sendBookingConfirmationEmail } from "./send-mail";